requirements.txt
lightgbm==4.5.0
fastapi==0.115.6
pandas==2.2.3
numpy==2.0.2
python-dotenv==1.0.1
uvicorn==0.34.0
boto3==1.35.93